+++ test/example/parentAction/foo/bar/view.js
@@ -1,5 +1,4 @@
 module.exports = {
-    parentAction: 'view',
     params: 'barId',
     get: function(req, res, next) {} // jshint ignore:line
 };