+++ test/example/parentAction/foo/index.js
@@ -1,4 +1,4 @@
 module.exports = {
-    parentAction: 'view',
+    parentAction: '../view',
     get: function(req, res, next) {} // jshint ignore:line
 };