+++ test/example/path/alter/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-    get: function(req, res, next) {} // jshint ignore:line
-};