+++ test/example/pathPart/foo/inherit.js
@@ -1,4 +1,4 @@
 module.exports = {
-    parentAction: 'edit',
+    parentAction: '../edit',
     get: function(req, res, next) {} // jshint ignore:line
 };