+++ test/example/pathPart/removed/view.js
@@ -1,4 +1,3 @@
 module.exports = {
-    params: 'id',
     get: function(req, res, next) {} // jshint ignore:line
 };